# openapi-lexia-hub.yaml
openapi: 3.1.0
info:
  title: LexIA API
  version: 2.0.0
  description: API unificada para legislaciones y jurisprudencia (SJF) normalizada para GPT.
servers:
  - url: https://lexia-api.vercel.app
paths:
...
  /ley:
    get:
      operationId: buscarLey
      summary: Buscar leyes por id categoria o nombre
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: integer
        - name: categoria
          in: query
          required: false
          schema:
            type: integer
        - name: nombre
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Lista de leyes encontradas
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    categoria:
                      type: integer
                    nombre:
                      type: string
  /jurisprudencia/buscar:
    get:
      operationId: buscarJurisprudencia
      summary: Buscar jurisprudencia SJF por termino simple
      description: Si no se envia q devuelve resultados generales con filtros por defecto.
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          example: amparo
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - name: includeRaw
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Resultado paginado normalizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JurisprudenciaSearchResponse"
        "400":
          description: Parametros invalidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Error de conexion con SJF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: buscarJurisprudenciaAvanzada
      summary: Buscar jurisprudencia SJF con payload avanzado
      description: Envia el body completo del formato de busqueda SJF.
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - name: includeRaw
          in: query
          required: false
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                classifiers:
                  type: array
                  items:
                    type: object
                    properties: {}
                searchTerms:
                  type: array
                  items:
                    type: object
                    properties: {}
                bFacet:
                  type: boolean
                ius:
                  type: array
                  items:
                    type: integer
                idApp:
                  type: string
                lbSearch:
                  type: array
                  items:
                    type: string
                filterExpression:
                  type: string
      responses:
        "200":
          description: Resultado paginado normalizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JurisprudenciaSearchResponse"
        "400":
          description: Parametros invalidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Error de conexion con SJF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jurisprudencia/detalle:
    get:
      operationId: obtenerDetalleJurisprudencia
      summary: Obtener detalle de tesis por IUS
      description: Si no se envia isSemanal el proxy prueba true y luego false.
      parameters:
        - name: ius
          in: query
          required: true
          schema:
            type: integer
          example: 2031780
        - name: isSemanal
          in: query
          required: false
          schema:
            type: boolean
        - name: hostName
          in: query
          required: false
          schema:
            type: string
            default: https://sjf2.scjn.gob.mx
        - name: includeRaw
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Detalle normalizado de tesis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JurisprudenciaDetailResponse"
        "400":
          description: Parametros invalidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Error de conexion con SJF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jurislex/decretos:
    get:
      operationId: obtenerDecretosJurislex
      summary: Obtener anexos/decretos de una legislacion
      parameters:
        - name: idLegislacion
          in: query
          required: true
          schema:
            type: integer
        - name: idOrdenamiento
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Lista de decretos o anexos
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      type: object
                      properties: {}
        "502":
          description: Error de conexion con Jurislex
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jurislex/articulos/buscar:
    get:
      operationId: buscarArticulosJurislex
      summary: Buscar articulos en Jurislex
      description: Busqueda por numero o palabra clave usando categoria e idLegislacion validados.
      parameters:
        - name: categoria
          in: query
          required: true
          schema:
            type: integer
        - name: idLegislacion
          in: query
          required: true
          schema:
            type: integer
        - name: desc
          in: query
          required: false
          schema:
            type: string
            default: ""
        - name: soloArticulo
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: indice
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: elementos
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: articuloNumero
          in: query
          required: false
          schema:
            type: integer
        - name: includeRaw
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Resultado paginado normalizado de articulos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JurislexSearchResponse"
        "400":
          description: Parametros invalidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Error de conexion con Jurislex
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: buscarArticulosJurislexAvanzado
      summary: Buscar articulos con payload avanzado Jurislex
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                categoria:
                  type: integer
                includeRaw:
                  type: boolean
                datosArticulo:
                  type: object
                  properties:
                    IdLegislacion:
                      type: array
                      items:
                        type: integer
                    Desc:
                      type: string
                    Indice:
                      type: integer
                    Elementos:
                      type: integer
                    Ordenamiento:
                      type: string
                    SoloArticulo:
                      type: boolean
                    SoloIndices:
                      type: boolean
                    filterRaw:
                      type: string
      responses:
        "200":
          description: Resultado paginado normalizado de articulos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JurislexSearchResponse"
        "400":
          description: Parametros invalidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Error de conexion con Jurislex
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jurislex/articulos/detalle:
    get:
      operationId: obtenerDetalleArticuloJurislex
      summary: Obtener texto completo de un articulo en Jurislex
      parameters:
        - name: categoria
          in: query
          required: true
          schema:
            type: integer
        - name: idLegislacion
          in: query
          required: true
          schema:
            type: integer
        - name: idArticulo
          in: query
          required: true
          schema:
            type: integer
        - name: includeRaw
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Detalle normalizado de articulo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JurislexDetailResponse"
        "400":
          description: Parametros invalidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Error de conexion con Jurislex
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    JurisprudenciaSearchResponse:
      type: object
      properties:
        total:
          type: integer
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer
        count:
          type: integer
        hasMore:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/JurisprudenciaItem"
      required:
        - total
        - totalPages
        - page
        - size
        - count
        - hasMore
        - items
    JurisprudenciaDetailResponse:
      type: object
      properties:
        ius:
          type: integer
        isSemanalUsed:
          type: boolean
        hostName:
          type: string
        rubro:
          type: string
        fechaPublicacion:
          type: string
        titulo:
          type: string
        texto:
          type: string
        textoPlano:
          type: string
        raw:
          type: object
          properties: {}
      required:
        - ius
        - isSemanalUsed
        - hostName
        - rubro
        - texto
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    JurislexSearchResponse:
      type: object
      properties:
        categoria:
          type: integer
        idLegislacion:
          type: integer
        indice:
          type: integer
        elementos:
          type: integer
        count:
          type: integer
        total:
          type: integer
        totalArticulos:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/JurislexItem"
      required:
        - categoria
        - count
        - total
        - totalArticulos
        - items
...